Issue #7: Capturing the first parameter to the install script to tell it a new hook was installed.
After a new hook is installed, they will "dial back" to git-guppy/scripts/install. Given that
the gulp execution is async, the git-guppy/scripts/install was reading the App's package.json
before it was updated. In order to overcome this, the install script will simply call
git-guppy/scripts/install "hook-name", adding the "hook-name" to the list of "hooks" to be
installed.

*	modified:   scripts/install.js
- Capturing the command line argument that's the name of the hook just installed.
- Adding it to the list of hooks
- The rest is history :)