Issue #7: Changing the status code to always 0 and outputing error, as Jenkins removes the module on 1.
This has become an issue on CI/CD environments, as the script returns 1 and that
makes npm to uninstall the script. For now, just returning 0 if it passes or not.

Adding a try/catch block to capture the exception nicely and output the error, but
continue with the install of other modules.