fix: ensure that Git commands will work in directories outside of the repository root
Git hooks pass several environment variables (prefixed with "GIT_") to be used by hook processes and Git commands executed from hooks. Two of these variables, GIT_DIR and GIT_INDEX_FILE, are set to the location and path of the root .git directory and .git/index file. Git commands executed from hooks use these environment variables as an implicit "--git-dir" argument, and if they don't refer to valid directories then those commands will fail.

Because Git executes hooks from the location of the .git directory, these two variables will be incorrectly set for guppy hooks if the directory containing the gulpfile is not the same as the repository root. If we detect that the value of GIT_DIR does not exist as a directory from our current working directory, then we remove both GIT_DIR and GIT_INDEX_FILE from the environment so that subsequent Git commands will autodetect the repository root and function properly.